Updated kernels, working 4.13 boot
Board was not booting using provided uboot so now using uboot from
meta-sunxi. Then erazor kernel was updated but hanged on boot after a
few lines. Went to megous kernel on 4.9 that booted successfully but did
not have working Ethernet after updating upstream. Finally created 4.13
megous recipe that booted and with working Ethernet (but strange
Ethernet led handling) and heartbeat led that was not in the 4.9 dts.